chaned to longs for higher precision

+++ obliv_math_func.h
@@ -10,21 +10,21 @@ typedef obliv int oint;
 
 #define DEBUG(...)
 #else
-typedef int oint;
+typedef long oint;
 #include <stdio.h>
 #define DEBUG(...) fprintf (stderr, __VA_ARGS__)
 
 #endif
 
 static inline oint mult_oo(oint x, oint y) {
-  return (x*y) / (1 << PRECISION);
+  return (x*y) / (1L << PRECISION);
 }
 
-static inline oint mult_op(oint x, int y) {
-  return (x*y) / (1 << PRECISION);
+static inline oint mult_op(oint x, long y) {
+  return (x*y) / (1L << PRECISION);
 }
 
-static inline void mult_ovec_p(oint* vec, int p, int num_entries) {
+static inline void mult_ovec_p(oint* vec, long p, int num_entries) {
   for (int i = 0; i < num_entries; i++) {
     vec[i] = mult_op(vec[i], p);
   }
@@ -36,7 +36,7 @@ static inline void add_ovecs(oint* vec1, oint* vec2, int num_entries) {
   }
 }
 
-static inline oint evalPolynomial(oint x, int* coeffs, int degree) {
+static inline oint evalPolynomial(oint x, long* coeffs, int degree) {
   // assume that the first coefficient in the constant term, the second is the linear term etc.
   oint val = coeffs[degree];
 
@@ -60,6 +60,7 @@ static inline oint oblivious_logistic_fn(oint* model, oint* features, int num_fe
 
 static inline void add_to_gradient(oint* grad, oint* model, oint* features, oint label, int num_features) {
   oint err = oblivious_logistic_fn(model, features, num_features) - (label << PRECISION);
+
   for (int k = 0; k < num_features; k++) {
     grad[k] += mult_oo(err, features[k]);
   }
